Load lib target name from `cargo metadata`.
