+++ src/manifest.rs
@@ -44,32 +44,6 @@ pub(crate) fn get_package_version(manifest: &Manifest) -> anyhow::Result<&str> {
     Ok(version)
 }
 
-pub(crate) fn get_lib_target_name(manifest: &Manifest) -> anyhow::Result<String> {
-    // If there's a [lib] section, return the name it specifies, if any.
-    if let Some(product) = &manifest.parsed.lib {
-        if let Some(lib_name) = &product.name {
-            return Ok(lib_name.clone());
-        }
-    }
-
-    // Otherwise, assume the crate is a lib crate with the default lib target name:
-    // the same name as the package but with dashes replaced with underscores.
-    Ok(get_package_name(manifest)?.replace('-', "_"))
-}
-
-pub(crate) fn get_first_bin_target_name(manifest: &Manifest) -> anyhow::Result<String> {
-    // If there's a [[bin]] section, return the first item's name.
-    if let Some(product) = manifest.parsed.bin.first() {
-        if let Some(bin_name) = &product.name {
-            return Ok(bin_name.clone());
-        }
-    }
-
-    // Otherwise, assume the crate is a bin crate with the default bin target name:
-    // the same name as the package but with dashes replaced with underscores.
-    Ok(get_package_name(manifest)?.replace('-', "_"))
-}
-
 pub(crate) fn get_project_dir_from_manifest_path(
     manifest_path: &std::path::Path,
 ) -> anyhow::Result<std::path::PathBuf> {
